requirements.txt
# Core Framework
fastapi>=0.95.0
uvicorn[standard]>=0.21.0
pydantic>=2.0.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0

# Data Sources
pytdx>=1.72
akshare>=1.12.0

# Utilities
pyyaml>=6.0
python-dateutil>=2.8.2
pytz>=2023.3
pandas>=2.0.0
numpy>=1.24.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Optional: For better profiling
memory-profiler>=0.61.0
